NI-2639 Fix invalid path for local lib in simple server
This webservice was originally copied from the rest webservice
which is built into Moodle. When it was being copied, some of the
paths were not updated and so still point to the location of the
original webservices's library. This commit fixers an invalid path
and also updates the check for ensuring the webservice is enabled.